Python (receive_ack.py): apply fair dispatch
Until now, RabbitMQ would send jobs in a 'Round Robin' fashion which
blindly distributes messages to consumers no matter of their workload.

With this commit, I've set the 'prefetch_count=1' in the 'quality of
service' attribute of the 'channel' so that RabbitMQ takes note of the
number of jobs in each consumer to distribute the messages to idle
consumers first!